Projeto-backendII/requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
sqlalchemy>=2.0.20
crewai>=0.28.1
langchain>=0.0.312
pydantic>=2.3.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
pytest>=7.4.2
httpx>=0.24.1
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
langchain-community>=0.0.1
langchain-openai>=0.1.1
duckduckgo-search>=0.0.1
email-validator>=2.0.0
requests>=2.31.0
langchain-ollama>=0.0.1
newspaper3k>=0.2.8
beautifulsoup4>=4.12.2
aiosqlite>=0.19.0
asyncpg>=0.28.0
psycopg2-binary>=2.9.9
prometheus-fastapi-instrumentator>=6.1.0
slowapi>=0.1.8